+++ config/importmap.rb
@@ -4,6 +4,7 @@ pin "cafe_car"
 pin "@hotwired/turbo-rails", to: "turbo.min.js"
 pin "@rails/actiontext", to: "actiontext.esm.js"
 pin "trix"
+pin "tom-select", to: "tom-select.complete.min.js" # searchable association selects
 
 # pin "@hotwired/stimulus", to: "stimulus.min.js"
 # pin "@hotwired/stimulus-loading", to: "stimulus-loading.js"